+++ cloud-siem-20241212/src/Models/ListResponseRulesResponseBody/responseRules.php
@@ -44,6 +44,11 @@ class responseRules extends Model
     public $responseRulePriority;
 
     /**
+     * @var string
+     */
+    public $responseRuleRemark;
+
+    /**
      * @var int
      */
     public $responseRuleStatus;
@@ -70,6 +75,7 @@ class responseRules extends Model
         'responseRuleId' => 'ResponseRuleId',
         'responseRuleName' => 'ResponseRuleName',
         'responseRulePriority' => 'ResponseRulePriority',
+        'responseRuleRemark' => 'ResponseRuleRemark',
         'responseRuleStatus' => 'ResponseRuleStatus',
         'responseRuleType' => 'ResponseRuleType',
         'responseTriggerType' => 'ResponseTriggerType',
@@ -110,6 +116,10 @@ class responseRules extends Model
 
         if (null !== $this->responseRulePriority) {
             $res['ResponseRulePriority'] = $this->responseRulePriority;
+        }
+
+        if (null !== $this->responseRuleRemark) {
+            $res['ResponseRuleRemark'] = $this->responseRuleRemark;
         }
 
         if (null !== $this->responseRuleStatus) {
@@ -165,6 +175,10 @@ class responseRules extends Model
 
         if (isset($map['ResponseRulePriority'])) {
             $model->responseRulePriority = $map['ResponseRulePriority'];
+        }
+
+        if (isset($map['ResponseRuleRemark'])) {
+            $model->responseRuleRemark = $map['ResponseRuleRemark'];
         }
 
         if (isset($map['ResponseRuleStatus'])) {